requirements.txt
amqp==2.6.0
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.6
click==7.1.2
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
redis==3.5.3
six==1.15.0
vine==1.3.0
wcwidth==0.2.5
Werkzeug==1.0.1
